Python: Share Foundry request context helpers
Move protocol validation and user-scope derivation into a dedicated request-context module, leaving the session-store module focused on storage.


Copilot-Session: 3e5c81ad-75e8-4e92-a883-8bbd676c6c8c
